Moving encrypt, sign and send logic from gateway (SMTP) to a MailService
